Implement a function to customize displayed values
I said that it is impossible to return a type because we don't have
types. I was dead wrong. This is Python! You can do whatever you
like.
1. You can create types on runtime.
2. There is a function called `sys.displayhook` which is
responsible for displaying things in interactive environment.
So that means we can return Python types and modify how they will
be printed. This will make `type(5)` return an `int` but it will
be displayed as `<sınıf 'tamsayı'>` in interactive console.